@@ -8,6 +8,7 @@ import static org.junit.Assert.fail;
 import java.util.List;
 import java.util.Vector;
 
+import javafx.application.Platform;
 import javafx.geometry.VerticalDirection;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
@@ -21,6 +22,7 @@ import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.loadui.testfx.GuiTest;
 import org.loadui.testfx.categories.TestFX;
+import org.loadui.testfx.utils.FXTestUtils;
 
 import rmblworx.tools.timey.gui.DateTimeUtil;
 
@@ -104,7 +106,12 @@ public class TimePickerTest extends GuiTest {
 
 		// Wert des Sliders setzen und sicherstellen, dass Textfeld-Inhalt der Erwartung entspricht
 		for (final TextFieldAndSliderValue testCase : testCases) {
-			testCase.slider.setValue(testCase.sliderValue);
+			Platform.runLater(new Runnable() {
+				public void run() {
+					testCase.slider.setValue(testCase.sliderValue);
+				}
+			});
+			FXTestUtils.awaitEvents();
 			try {
 				waitUntil(testCase.textField, new Predicate<TextField>() {
 					public boolean apply(final TextField textField) {
@@ -119,7 +126,12 @@ public class TimePickerTest extends GuiTest {
 
 		// Inhalt des Textfelds setzen und sicherstellen, dass Slider-Wert der Erwartung entspricht
 		for (final TextFieldAndSliderValue testCase : testCases) {
-			testCase.textField.setText(testCase.textFieldValue);
+			Platform.runLater(new Runnable() {
+				public void run() {
+					testCase.textField.setText(testCase.textFieldValue);
+				}
+			});
+			FXTestUtils.awaitEvents();
 			try {
 				waitUntil(testCase.slider, new Predicate<Slider>() {
 					public boolean apply(final Slider slider) {
@@ -146,7 +158,7 @@ public class TimePickerTest extends GuiTest {
 
 		for (final TextFieldScrollingData testCase : testCases) {
 			click(testCase.focusedTextField); // anderes Feld fokussieren
-			// Scrollen über Feld ohne Fokus hat keine Auswirkung 
+			// Scrollen über Feld ohne Fokus hat keine Auswirkung
 			move(testCase.textField);
 			scroll(VerticalDirection.UP);
 			assertEquals(0.0, testCase.slider.getValue(), 0.0);
@@ -188,14 +200,24 @@ public class TimePickerTest extends GuiTest {
 		for (final InputMode inputMode : new InputMode[] { InputMode.TYPE, InputMode.COPY_PASTE }) {
 			// Inhalt des Textfelds setzen und sicherstellen, dass Inhalt der Erwartung entspricht
 			for (final TextFieldInputAndValue testCase : testCases) {
-				testCase.textField.requestFocus(); // Feld fokussieren
+				Platform.runLater(new Runnable() {
+					public void run() {
+						testCase.textField.requestFocus(); // Feld fokussieren
+					}
+				});
+				FXTestUtils.awaitEvents();
 				assertTrue(testCase.textField.isFocused());
 				switch (inputMode) {
 					case TYPE:
 						type(testCase.input);
 						break;
 					case COPY_PASTE:
-						testCase.textField.setText(testCase.input);
+						Platform.runLater(new Runnable() {
+							public void run() {
+								testCase.textField.setText(testCase.input);
+							}
+						});
+						FXTestUtils.awaitEvents();
 						break;
 					default:
 						fail("unbekannter Modus");
@@ -215,7 +237,12 @@ public class TimePickerTest extends GuiTest {
 							inputMode.toString().toLowerCase(), testCase.textField.getText(), testCase.value));
 				} finally {
 					// Feldinhalt auf Ausgangswert zurücksetzen
-					testCase.textField.setText("00");
+					Platform.runLater(new Runnable() {
+						public void run() {
+							testCase.textField.setText("00");
+						}
+					});
+					FXTestUtils.awaitEvents();
 				}
 			}
 		}
@@ -233,7 +260,12 @@ public class TimePickerTest extends GuiTest {
 
 		for (final TimePartsAndTimeValue testCase : testCases) {
 			final LocalTime time = DateTimeUtil.getLocalTimeForString(testCase.timeString);
-			timePicker.setTime(time);
+			Platform.runLater(new Runnable() {
+				public void run() {
+					timePicker.setTime(time);
+				}
+			});
+			FXTestUtils.awaitEvents();
 			assertEquals(testCase.hours, hoursTextField.getText());
 			assertEquals(testCase.minutes, minutesTextField.getText());
 			assertEquals(testCase.seconds, secondsTextField.getText());
@@ -253,9 +285,14 @@ public class TimePickerTest extends GuiTest {
 
 		for (final TimePartsAndTimeValue testCase : testCases) {
 			final LocalTime time = DateTimeUtil.getLocalTimeForString(testCase.timeString);
-			hoursTextField.setText(testCase.hours);
-			minutesTextField.setText(testCase.minutes);
-			secondsTextField.setText(testCase.seconds);
+			Platform.runLater(new Runnable() {
+				public void run() {
+					hoursTextField.setText(testCase.hours);
+					minutesTextField.setText(testCase.minutes);
+					secondsTextField.setText(testCase.seconds);
+				}
+			});
+			FXTestUtils.awaitEvents();
 			assertEquals(time, timePicker.getTime());
 			assertEquals(time, timePicker.getTimeProperty().get());
 		}
